move get browser locale
